Make editor a rollup package
