+++ qt/aqt/data/web/js/BUILD.bazel
@@ -1,9 +1,12 @@
 load("@npm//@bazel/typescript:index.bzl", "ts_library")
+load("@npm//@bazel/rollup:index.bzl", "rollup_bundle")
 load("//ts:prettier.bzl", "prettier_test")
+load("//ts:eslint.bzl", "eslint_test")
 
 ts_library(
     name = "pycmd",
     srcs = ["pycmd.d.ts"],
+    visibility = ["//qt/aqt/data/web/js:__subpackages__"],
 )
 
 ts_library(
@@ -26,10 +29,30 @@ filegroup(
     output_group = "es5_sources",
 )
 
+###### aqt bundles
+
+rollup_bundle(
+    name = "editor",
+    config_file = "rollup.config.js",
+    entry_point = "//qt/aqt/data/web/js/editor:index.ts",
+    format = "iife",
+    link_workspace_root = True,
+    silent = True,
+    sourcemap = "false",
+    deps = [
+        "//qt/aqt/data/web/js/editor",
+        "@npm//@rollup/plugin-commonjs",
+        "@npm//@rollup/plugin-node-resolve",
+        "@npm//rollup-plugin-terser",
+    ],
+)
+
+
 filegroup(
     name = "js",
     srcs = [
         "aqt_es5",
+        "editor",
         "mathjax.js",
         "//qt/aqt/data/web/js/vendor",
     ],
@@ -47,4 +70,7 @@ prettier_test(
 #     srcs = glob(["*.ts"]),
 # )
 
-exports_files(["mathjax.js"])
+exports_files([
+    "mathjax.js",
+    "tsconfig.json",
+])
