+++ qt/aqt/data/web/js/editor/BUILD.bazel
@@ -1,0 +1,16 @@
+load("@npm//@bazel/typescript:index.bzl", "ts_library")
+
+ts_library(
+    name = "editor",
+    srcs = glob(["*.ts"]),
+    tsconfig = "//qt/aqt/data/web/js:tsconfig.json",
+    deps = [
+        "//qt/aqt/data/web/js:pycmd",
+        "@npm//@types/jquery",
+    ],
+    visibility = ["//qt:__subpackages__"],
+)
+
+exports_files([
+    "index.ts",
+])
